fix: fix infinite loop when server throws 500 on token check
