+++ src/core/clients/client.ts
@@ -6,7 +6,7 @@ import type {
   RegisterInterceptorsType,
 } from "./types";
 
-const createClient = (baseUrl: string, api = fetch) => {
+export const createClient = (baseUrl: string, api = fetch) => {
   let interceptRequest = (url: string, options: RequestInit) => {
     return { interceptedUrl: url, interceptedOptions: options };
   };
@@ -103,56 +103,3 @@ const createClient = (baseUrl: string, api = fetch) => {
     },
   };
 };
-
-const apiUrl = process.env.REACT_APP_SERVER_URL;
-if (!apiUrl) {
-  throw new Error("apiUrl environment variable not set");
-}
-const appApi = createClient(apiUrl, fetch);
-
-appApi.registerInterceptors({
-  response: async (response: Response) => {
-    if (isInternalServerError(response.status)) {
-      window.location.assign("/error/500");
-    }
-    if (isForbidden(response.status)) {
-      window.location.assign("/login");
-    }
-    if (isBadRequest(response.status)) {
-      const exceptionData: BadRequestError = await response.json();
-      throw exceptionData;
-    }
-    if (isUnauthorized(response.status)) {
-      const exceptionData: UnauthorizedError = await response.json();
-      throw exceptionData;
-    }
-    if (isOtherError(response.status)) {
-      throw "Other error: " + response.statusText;
-    }
-    return response;
-  },
-  responseError: (error: Error) => {
-    window.location.assign("/error/500");
-  },
-});
-
-const isInternalServerError = (statusCode: number) => {
-  return statusCode === 500;
-};
-
-const isForbidden = (statusCode: number) => {
-  return statusCode === 403;
-};
-
-const isBadRequest = (statusCode: number) => {
-  return statusCode === 400;
-};
-const isUnauthorized = (statusCode: number) => {
-  return statusCode === 401;
-};
-
-const isOtherError = (statusCode: number) => {
-  return statusCode >= 400;
-};
-
-export { appApi };
