+++ src/core/services/user.ts
@@ -1,5 +1,5 @@
 import { User } from "../../types/globalTypes";
-import { appApi } from "../clients/appApi";
+import { noRedirectApi, primaryApi } from "../clients/apis";
 
 interface Credentials {
   email: string;
@@ -11,22 +11,24 @@ interface RegisterValues extends Credentials {
 }
 
 export async function signIn(body: Credentials) {
-  return await appApi.post("users/signin", body, {
+  return await primaryApi.post("users/signin", body, {
     credentials: "include",
   });
 }
 
 export async function signOut() {
-  return await appApi.post("users/signout", {}, { credentials: "include" });
+  return await primaryApi.post("users/signout", {}, { credentials: "include" });
 }
 
 export async function register(body: RegisterValues) {
-  return await appApi.post("users/register", body, {
+  return await primaryApi.post("users/register", body, {
     credentials: "include",
   });
 }
 
 export async function checkToken() {
-  const res = await appApi.get("users/check-token", { credentials: "include" });
+  const res = await noRedirectApi.get("users/check-token", {
+    credentials: "include",
+  });
   return (await res.json()) as User;
 }
